ENH: Update Binder to work with PyQt5.
PyQt5 removed SIGNAL(), which was a useful way to represent an unbound
Qt signal in all of its possibly-overloaded glory. Instead, we have to
work with the bound signal object and navigate to the correct overload.

This does cause a minor behavior change in the case where the overload
is due to a default value, i.e. when we have an argumentless overload
and a single-argument overload. Through the bound signal, we can only
access the single-argument version. This is a behavior change, but the
new behavior is actually usually what we wanted anyways, so we let it
slide after adjusting the test suite.